updated and working

+++ main.py
@@ -13,9 +13,8 @@ import utils
 import waif 
 from RadioProgram import RadioProgram
 
-DEBUG = True
+DEBUG = False
 
-BUFFER_FILE = "threadbuffer.json"
 PROGRAM_CSV_FILE = "Shows.csv"
 
 if DEBUG:
@@ -27,10 +26,6 @@ else:
     ARCHIVE_DIRECTORY = os.path.join("W:", os.sep, "OneDrive - The Real StepChild Radio Of Cincinnati One", "New Recording")
     CHECK_FILE_DURATION_SECONDS = 5*60
     UPLOAD_WAIT_HOURS = 1
-    
-    
-ARCHIVE_DIRECTORY = "example"
-PROGRAM_CSV_FILE = "Shows.csv"
 
 def match_archive_to_program(archive_datetime: datetime.datetime, programs: list):
     # find program 
@@ -56,7 +51,8 @@ def main():
     
     # set up logging
     log_path = "log"
-    os.remove(log_path)
+    if os.path.exists(log_path):
+        os.remove(log_path)
     logging.basicConfig(level=logging.INFO, 
                     format="%(asctime)s [%(levelname)s]: %(name)s: %(message)s", 
                     handlers=[
